refactor: 우선순위 api호출을 한번 하도록 개선
