+++ src/components/App.jsx
@@ -10,6 +10,7 @@ import Login from "../pages/Login";
 import MyPage from "./MyPage";
 import Nav from "../Nav";
 import PasswordError from "../error/PasswordError";
+import { PriorityProvider } from "../contexts/PriorityContext";
 import React from "react";
 import { TodayLectureProvider } from "../contexts/TodayLectureContext";
 import { UserSettingsProvider } from "../contexts/UserSettingsContext";
@@ -18,22 +19,24 @@ const App = () => {
   return (
     <UserSettingsProvider>
       <TodayLectureProvider>
-        <Nav />
-        <Routes>
-          <Route path="/home" element={<Home />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/passwordError" element={<PasswordError />} />
-          <Route path="/courses/:courseCode" element={<Courses />} />
-          <Route
-            path="/courses/:courseCode/notices"
-            element={<CourseNotice />}
-          />
-          <Route path="/courses" element={<CoursesList />} />
-          <Route path="/mypage" element={<MyPage />} />
-          <Route path="/about" element={<About />} />
-          <Route path="/LoginError" element={<AuthErrorPage />} />
-          <Route path="*" element={<Navigate to="/not-found" replace />} />
-        </Routes>
+        <PriorityProvider>
+          <Nav />
+          <Routes>
+            <Route path="/home" element={<Home />} />
+            <Route path="/login" element={<Login />} />
+            <Route path="/passwordError" element={<PasswordError />} />
+            <Route path="/courses/:courseCode" element={<Courses />} />
+            <Route
+              path="/courses/:courseCode/notices"
+              element={<CourseNotice />}
+            />
+            <Route path="/courses" element={<CoursesList />} />
+            <Route path="/mypage" element={<MyPage />} />
+            <Route path="/about" element={<About />} />
+            <Route path="/LoginError" element={<AuthErrorPage />} />
+            <Route path="*" element={<Navigate to="/not-found" replace />} />
+          </Routes>
+        </PriorityProvider>
       </TodayLectureProvider>
     </UserSettingsProvider>
   );
