+++ src/contexts/TodayLectureContext.js
@@ -27,7 +27,6 @@ export const TodayLectureProvider = ({ children }) => {
         setIsLoading(true);
         console.log("오늘의 강의데이터 호출");
         const data = await fetchMemberTodayLecture();
-        console.log("API 응답 데이터:", data);
         setTodayLecture(data?.contents ? data : { contents: [] });
       } catch (error) {
         console.error("오늘의 강의 조회 실패:", error);
